Move the samples to their own project and directory.
