+++ Samples/iOS/Images.xcassets/AppIcon.appiconset/Contents.json
