+++ Samples/iOS/Info.plist
