+++ Samples/macOS/Assets.xcassets/AppIcon.appiconset/Contents.json
