+++ Samples/macOS/Info.plist
