+++ Samples/tvOS/Assets.xcassets/App Icon & Top Shelf Image.brandassets/Contents.json
@@ -13,6 +13,12 @@
       "role" : "primary-app-icon"
     },
     {
+      "size" : "2320x720",
+      "idiom" : "tv",
+      "filename" : "Top Shelf Image Wide.imageset",
+      "role" : "top-shelf-image-wide"
+    },
+    {
       "size" : "1920x720",
       "idiom" : "tv",
       "filename" : "Top Shelf Image.imageset",
