+++ Samples/tvOS/Assets.xcassets/Contents.json
