+++ Samples/tvOS/Assets.xcassets/LaunchImage.launchimage/Contents.json
